Add display of backtrace when there's a runtime error
The implementation for this introduces a performance regression.  This
commit changes the interpreter to actually create a separate call stack
with the sole purpose of being able to print the backtrace if a runtime
error occurs.  Why?  Because we actually rely on Rust's call stack for
the real call stack.  This reifies the call stack so that we can inspect
it at will.

In the future, Environments and CallFrames will probably be combined
somehow.