Fix frame, scope, and slot terminology to be clearer
The main change here is renaming FrameIndex to SlotIndex.  The term
FrameIndex was never clear about whether it was the index *of* a frame
or the index *in* a frame.

Also up until now, we've conflated the terms "frame" and "scope".
Scope is purely lexical.  Frame is based on runtime execution flow.
"Frame" is shorthand for call frame, which heavily implies the stack.
"Stackframe" and "callframe" are basically synonymous.  However, some
language implementations do put call frames on the heap.

The term "slot" is agnostic about whether it's a slot in the stack or a
slot on the heap.  Since we may move local variables to a stack for
better runtime performance in the future, I thought that an agnostic
term was a good choice.