Change EnvironmentRef to be an index into an arena
Instead of linked lists of environments, we use an arena, which is
essentially an owned vector, with extra runtime safety guarantees when
accessing elements.  An environment is now an index into the arena.  It
has an optional enclosing index (i.e. the thing that used to be the
parent pointer in the linked list).  This allows us to remove all
RefCells from runtime environments.  The Interpreter now owns the
entire arena of all environments.

First, the good news... Runtime performance is much faster.  The
benchmark runs 20% faster!  This is huge.  But there is some bad news.

A side-effect of using indexes into an arena is that we removed the
reference counting of environments.  This means that Rc no longer
deallocates environments for us.  So currently, we never free
environments, which is a significant memory leak.  But since there
were almost certainly cycles in the references, reference counting was
never the right approach in the first place.

We are going to need to implement a garbage collector, or as I like to
call it, a memory recycler.